+++ src/panels/lovelace/hui-view.js
@@ -1,17 +1,17 @@
 import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
-import './hui-camera-preview-card.js';
-import './hui-entities-card.js';
-import './hui-entity-filter-card.js';
-import './hui-glance-card';
-import './hui-history-graph-card.js';
-import './hui-media-control-card.js';
-import './hui-picture-card.js';
-import './hui-picture-glance-card';
-import './hui-plant-status-card.js';
-import './hui-weather-forecast-card';
-import './hui-error-card.js';
+import './cards/hui-camera-preview-card.js';
+import './cards/hui-entities-card.js';
+import './cards/hui-entity-filter-card.js';
+import './cards/hui-glance-card';
+import './cards/hui-history-graph-card.js';
+import './cards/hui-media-control-card.js';
+import './cards/hui-entity-picture-card.js';
+import './cards/hui-picture-glance-card';
+import './cards/hui-plant-status-card.js';
+import './cards/hui-weather-forecast-card';
+import './cards/hui-error-card.js';
 
 import applyThemesOnElement from '../../common/dom/apply_themes_on_element.js';
 import computeCardElement from './common/compute-card-element.js';
